fixed duplicate friends posting
