+++ api/database.json
@@ -81,20 +81,14 @@
   ],
   "friends": [
     {
-      "friendInitiateId": 1,
-      "userId": 2,
-      "active": true,
-      "id": 1
-    },
-    {
       "friendInitiateId": 2,
       "userId": 3,
       "active": true,
       "id": 2
     },
     {
-      "friendInitiateId": 2,
-      "userId": 1,
+      "friendInitiateId": 1,
+      "userId": 2,
       "active": true,
       "id": 3
     },
@@ -102,7 +96,7 @@
       "friendInitiateId": 1,
       "userId": 3,
       "active": true,
-      "id": 4
+      "id": 5
     }
   ]
 }
