+++ src/scripts/friends/FriendCard.js
@@ -19,6 +19,7 @@ eventHub.addEventListener("click", clickEvent => {
 
    deleteFriends(friendId).then(() => {
      const message = new CustomEvent ("friendDeleted")
+     console.log("friend deleted")
      eventHub.dispatchEvent(message)
    })
  }
