+++ src/scripts/friends/FriendsList.js
@@ -1,8 +1,19 @@
 import { FriendCard } from "./FriendCard.js";
-import { useFriends } from "./FriendsProvider.js";
+import { useFriends, SaveFriends } from "./FriendsProvider.js";
 
 const eventHub = document.querySelector(".container");
 const contentElement = document.querySelector(".friendsCards");
+
+eventHub.addEventListener("newFriend", event => {
+  SaveFriends(event.detail).then(() => {
+    const updatedFriends = useFriends();
+    const activeUserId = parseInt(
+      sessionStorage.getItem("activeUser"), 10);
+    const updatedFoundFriendsArray = updatedFriends.filter(
+      friendRel => friendRel.friendInitiateId === activeUserId
+    );
+    render(updatedFoundFriendsArray);
+  })})
 
 export const FriendsListComponent = () => {
   
@@ -15,6 +26,12 @@ export const FriendsListComponent = () => {
 
       render(foundFriendsArray);
 
+      }
+      //custom event that says a new friend should be added to friends table
+
+     
+
+
       eventHub.addEventListener("friendDeleted", () => {
         console.log("deleted friend event heard");
         const updatedFriends = useFriends();
@@ -25,7 +42,7 @@ export const FriendsListComponent = () => {
         contentElement.innerHTML = "";
         render(updatedFoundFriendsArray);
       });
-    }
+    
   
 
   export const render = foundFriendsArray => {
