+++ src/scripts/messages/MessageListener.js
@@ -1,5 +1,4 @@
 import { MessageList } from "./MessageList.js";
-import { initiateFriendSearchComponent } from "../friends/FriendSearch.js";
 import { FriendsListComponent } from "../friends/FriendsList.js";
 import { addSearchEventListeners } from "../friends/FriendSearchList.js";
 import { NavbarEventListener } from "../navbar/navbarList.js";
